fix template

+++ test/ReputationRegistry.t.sol
@@ -11,12 +11,12 @@ import {IIdentityRegistry} from "../src/interfaces/IIdentityRegistry.sol";
 contract ReputationRegistryTest is Test {
     ReputationRegistry public registry;
     IdentityRegistry public identityRegistry;
-    
+
     uint256 public clientAgentId;
     uint256 public serverAgentId;
     uint256 public anotherClientAgentId;
     uint256 public anotherServerAgentId;
-    
+
     address public clientAddress = address(0x1);
     address public serverAddress = address(0x2);
     address public anotherClientAddress = address(0x3);
@@ -25,26 +25,26 @@ contract ReputationRegistryTest is Test {
     function setUp() public {
         // Deploy IdentityRegistry first
         identityRegistry = new IdentityRegistry();
-        
+
         // Deploy ReputationRegistry with IdentityRegistry address
         registry = new ReputationRegistry(identityRegistry);
-        
+
         // Register agents with proper roles
         vm.prank(clientAddress);
         clientAgentId = identityRegistry.newAgent("client.example.com", clientAddress);
         vm.prank(clientAddress);
         identityRegistry.addRole(clientAgentId, IIdentityRegistry.Role.CLIENT);
-        
+
         vm.prank(serverAddress);
         serverAgentId = identityRegistry.newAgent("server.example.com", serverAddress);
         vm.prank(serverAddress);
         identityRegistry.addRole(serverAgentId, IIdentityRegistry.Role.SERVER);
-        
+
         vm.prank(anotherClientAddress);
         anotherClientAgentId = identityRegistry.newAgent("client2.example.com", anotherClientAddress);
         vm.prank(anotherClientAddress);
         identityRegistry.addRole(anotherClientAgentId, IIdentityRegistry.Role.CLIENT);
-        
+
         vm.prank(anotherServerAddress);
         anotherServerAgentId = identityRegistry.newAgent("server2.example.com", anotherServerAddress);
         vm.prank(anotherServerAddress);
@@ -54,7 +54,7 @@ contract ReputationRegistryTest is Test {
     function test_AcceptFeedback() public {
         vm.prank(serverAddress); // Server owner authorizes feedback
         registry.acceptFeedback(clientAgentId, serverAgentId);
-        
+
         // No direct way to verify internal state, but function should execute successfully
         assertTrue(true);
     }
@@ -63,7 +63,7 @@ contract ReputationRegistryTest is Test {
         // We can't predict the exact feedbackAuthId, but we can verify the event structure
         vm.expectEmit(true, true, false, false);
         emit IReputationRegistry.AuthFeedback(clientAgentId, serverAgentId, bytes32(0));
-        
+
         vm.prank(serverAddress); // Server owner authorizes feedback
         registry.acceptFeedback(clientAgentId, serverAgentId);
     }
@@ -71,28 +71,28 @@ contract ReputationRegistryTest is Test {
     function test_AcceptFeedback_GeneratesUniqueFeedbackAuthId() public {
         // Record events to check feedbackAuthId uniqueness
         vm.recordLogs();
-        
+
         vm.prank(serverAddress);
         registry.acceptFeedback(clientAgentId, serverAgentId);
         vm.prank(serverAddress);
         registry.acceptFeedback(clientAgentId, serverAgentId);
-        
+
         Vm.Log[] memory logs = vm.getRecordedLogs();
-        
+
         // Should have 2 AuthFeedback events
         assertEq(logs.length, 2);
-        
+
         // Extract feedbackAuthIds from the events
         bytes32 feedbackAuthId1 = logs[0].topics[3];
         bytes32 feedbackAuthId2 = logs[1].topics[3];
-        
+
         // Should be different
         assertTrue(feedbackAuthId1 != feedbackAuthId2);
     }
 
     function test_AcceptFeedback_MultipleAgentCombinations() public {
         vm.recordLogs();
-        
+
         vm.prank(serverAddress);
         registry.acceptFeedback(clientAgentId, serverAgentId);
         vm.prank(serverAddress);
@@ -101,21 +101,21 @@ contract ReputationRegistryTest is Test {
         registry.acceptFeedback(clientAgentId, anotherServerAgentId);
         vm.prank(anotherServerAddress);
         registry.acceptFeedback(anotherClientAgentId, anotherServerAgentId);
-        
+
         Vm.Log[] memory logs = vm.getRecordedLogs();
-        
+
         // Should have 4 AuthFeedback events
         assertEq(logs.length, 4);
-        
+
         // Verify all events have different feedbackAuthIds
         bytes32[] memory authIds = new bytes32[](4);
-        for (uint i = 0; i < 4; i++) {
+        for (uint256 i = 0; i < 4; i++) {
             authIds[i] = logs[i].topics[3];
         }
-        
+
         // Check uniqueness
-        for (uint i = 0; i < 4; i++) {
-            for (uint j = i + 1; j < 4; j++) {
+        for (uint256 i = 0; i < 4; i++) {
+            for (uint256 j = i + 1; j < 4; j++) {
                 assertTrue(authIds[i] != authIds[j]);
             }
         }
@@ -124,20 +124,20 @@ contract ReputationRegistryTest is Test {
     function test_AcceptFeedback_CorrectEventParameters() public {
         vm.expectEmit(true, true, false, false);
         emit IReputationRegistry.AuthFeedback(clientAgentId, serverAgentId, bytes32(0));
-        
+
         vm.prank(serverAddress);
         registry.acceptFeedback(clientAgentId, serverAgentId);
     }
 
     function test_AcceptFeedback_FeedbackAuthIdIncludesChainId() public {
         vm.recordLogs();
-        
+
         vm.prank(serverAddress);
         registry.acceptFeedback(clientAgentId, serverAgentId);
-        
+
         Vm.Log[] memory logs = vm.getRecordedLogs();
         bytes32 feedbackAuthId = logs[0].topics[3];
-        
+
         // The feedbackAuthId should be deterministic based on chainid, contract address, and counter
         // We can't easily test the internal assembly logic, but we can verify it's not zero
         assertTrue(feedbackAuthId != bytes32(0));
@@ -146,48 +146,48 @@ contract ReputationRegistryTest is Test {
     function test_AcceptFeedback_FeedbackAuthIdIncludesContractAddress() public {
         // Deploy a second registry to compare feedbackAuthIds
         ReputationRegistry registry2 = new ReputationRegistry(identityRegistry);
-        
+
         vm.recordLogs();
-        
+
         vm.prank(serverAddress);
         registry.acceptFeedback(clientAgentId, serverAgentId);
         vm.prank(serverAddress);
         registry2.acceptFeedback(clientAgentId, serverAgentId);
-        
+
         Vm.Log[] memory logs = vm.getRecordedLogs();
-        
+
         // Should have 2 events
         assertEq(logs.length, 2);
-        
+
         bytes32 feedbackAuthId1 = logs[0].topics[3];
         bytes32 feedbackAuthId2 = logs[1].topics[3];
-        
+
         // Should be different because they come from different contract addresses
         assertTrue(feedbackAuthId1 != feedbackAuthId2);
     }
 
     function test_AcceptFeedback_IncrementalCounters() public {
         vm.recordLogs();
-        
+
         // Call multiple times to verify counter increments
-        for (uint i = 0; i < 5; i++) {
+        for (uint256 i = 0; i < 5; i++) {
             vm.prank(serverAddress);
             registry.acceptFeedback(clientAgentId, serverAgentId);
         }
-        
+
         Vm.Log[] memory logs = vm.getRecordedLogs();
-        
+
         // Should have 5 events, all with different feedbackAuthIds
         assertEq(logs.length, 5);
-        
+
         bytes32[] memory authIds = new bytes32[](5);
-        for (uint i = 0; i < 5; i++) {
+        for (uint256 i = 0; i < 5; i++) {
             authIds[i] = logs[i].topics[3];
         }
-        
+
         // Verify all are unique
-        for (uint i = 0; i < 5; i++) {
-            for (uint j = i + 1; j < 5; j++) {
+        for (uint256 i = 0; i < 5; i++) {
+            for (uint256 j = i + 1; j < 5; j++) {
                 assertTrue(authIds[i] != authIds[j]);
             }
         }
@@ -198,33 +198,33 @@ contract ReputationRegistryTest is Test {
         // Use our pre-configured agents instead
         vm.expectEmit(true, true, false, false);
         emit IReputationRegistry.AuthFeedback(clientAgentId, serverAgentId, bytes32(0));
-        
+
         vm.prank(serverAddress);
         registry.acceptFeedback(clientAgentId, serverAgentId);
     }
 
     function testFuzz_AcceptFeedback_MultipleCalls(uint8 numCalls) public {
         vm.assume(numCalls > 0 && numCalls <= 10); // Limit to avoid gas issues
-        
+
         vm.recordLogs();
-        
-        for (uint i = 0; i < numCalls; i++) {
+
+        for (uint256 i = 0; i < numCalls; i++) {
             vm.prank(serverAddress);
             registry.acceptFeedback(clientAgentId, serverAgentId);
         }
-        
+
         Vm.Log[] memory logs = vm.getRecordedLogs();
         assertEq(logs.length, numCalls);
-        
+
         // Verify all feedbackAuthIds are unique
         if (numCalls > 1) {
             bytes32[] memory authIds = new bytes32[](numCalls);
-            for (uint i = 0; i < numCalls; i++) {
+            for (uint256 i = 0; i < numCalls; i++) {
                 authIds[i] = logs[i].topics[3];
             }
-            
-            for (uint i = 0; i < numCalls; i++) {
-                for (uint j = i + 1; j < numCalls; j++) {
+
+            for (uint256 i = 0; i < numCalls; i++) {
+                for (uint256 j = i + 1; j < numCalls; j++) {
                     assertTrue(authIds[i] != authIds[j]);
                 }
             }
